+++ src/row/arch/neon/tests/mod.rs
@@ -7,6 +7,7 @@ mod packed_rgb;
 mod packed_rgb_float;
 mod packed_yuv_8bit;
 mod planar_8bit_and_nv;
+mod planar_gbr;
 mod v210;
 mod v30x;
 mod v410;