+++ src/sinker/mixed/tests/mod.rs
@@ -12,6 +12,7 @@ mod packed_rgb_10bit;
 mod packed_rgb_8bit;
 mod packed_rgb_float;
 mod packed_yuv_8bit;
+mod planar_gbr;
 mod planar_other_8bit_9bit;
 mod semi_planar_8bit;
 mod subsampled_4_2_0_high_bit;